+++ app/src/main/java/com/example/cammate/presentation/MainActivity.kt
@@ -13,7 +13,7 @@ import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.viewpager2.widget.ViewPager2
 import com.example.cammate.databinding.ActivityMainBinding
 import com.example.cammate.presentation.make_room.MakeRoomHostFragment
-import com.example.cammate.utils.setStatusBarTransparent
+import com.example.cammate.presentation.utils.setStatusBarTransparent
 import com.example.cammate.presentation.find_room.FindRoomHostFragment
 import com.example.cammate.webRTC.Models.MessageModel
 import com.example.cammate.webRTC.utils.NewMessageInterface
