+++ app/src/main/java/com/example/cammate/presentation/make_room/MakeRoomFragment.kt
@@ -2,6 +2,7 @@ package com.example.cammate.presentation.make_room
 
 import android.Manifest
 import android.content.Intent
+import android.content.res.Resources
 import android.os.Bundle
 import android.provider.Settings
 import android.util.Log
@@ -13,6 +14,7 @@ import android.widget.Toast
 import androidx.navigation.fragment.findNavController
 import com.example.cammate.R
 import com.example.cammate.databinding.FragmentMakeRoomBinding
+import com.example.cammate.presentation.chatting.ChatFragment
 import com.example.cammate.retrofit.PostRoom.PostRequest
 import com.example.cammate.retrofit.RetrofitWork
 import com.example.cammate.webRTC.Models.IceCandidateModel
@@ -21,6 +23,7 @@ import com.example.cammate.webRTC.RTCClient
 import com.example.cammate.webRTC.SocketRepository
 import com.example.cammate.webRTC.utils.NewMessageInterface
 import com.example.cammate.webRTC.utils.PeerConnectionObserver
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.permissionx.guolindev.PermissionX
 import org.webrtc.IceCandidate
 import org.webrtc.MediaStream
@@ -70,6 +73,12 @@ class MakeRoomFragment : Fragment(){
                     }
                 }
 
+            /*try {
+                val bottomsheet: BottomSheetDialogFragment = ChatFragment()
+                bottomsheet.show(requireFragmentManager(), "bottomsheet")
+            } catch (e: Exception){
+                Log.d("tag", "${e}")
+            }*/
         }
     }
