+++ app/src/main/java/com/example/cammate/presentation/utils/RandomName.kt
@@ -1,0 +1,4 @@
+package com.example.cammate.presentation.utils
+
+class RandomName {
+}
