+++ app/src/main/java/com/example/cammate/webRTC/SocketRespository.kt
@@ -21,7 +21,7 @@ class SocketRepository (private val messageInterface: NewMessageInterface) {
         // and get your ethernet ipv4 , mine is : "ws://192.168.1.3:3000"
         //but if your websocket is deployed you add your websocket address here
         //ws://13.125.228.93:3000
-        webSocket = object : WebSocketClient(URI("ws://10.0.2.2:3000")) {
+        webSocket = object : WebSocketClient(URI("ws://10.0.2.2:3000/")) {
             //        webSocket = object : WebSocketClient(URI("ws://192.168.1.3:3000")) {
             override fun onOpen(handshakedata: ServerHandshake?) {
                 sendMessageToSocket(
